Extracted xtext resource validation
