+++ subprojects/semantifyr/src/main/kotlin/reader/OxstsReader.kt
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2023-2024 The Semantifyr Authors
+ * SPDX-FileCopyrightText: 2023-2025 The Semantifyr Authors
  *
  * SPDX-License-Identifier: EPL-2.0
  */
@@ -9,22 +9,17 @@ package hu.bme.mit.semantifyr.oxsts.semantifyr.reader
 import hu.bme.mit.semantifyr.oxsts.lang.OxstsStandaloneSetup
 import hu.bme.mit.semantifyr.oxsts.model.oxsts.OxstsPackage
 import hu.bme.mit.semantifyr.oxsts.model.oxsts.Package
-import hu.bme.mit.semantifyr.oxsts.semantifyr.utils.error
+import hu.bme.mit.semantifyr.oxsts.semantifyr.utils.XtextResourceValidator
 import hu.bme.mit.semantifyr.oxsts.semantifyr.utils.info
 import hu.bme.mit.semantifyr.oxsts.semantifyr.utils.loggerFactory
-import hu.bme.mit.semantifyr.oxsts.semantifyr.utils.warn
+import hu.bme.mit.semantifyr.oxsts.semantifyr.utils.walkFiles
 import org.eclipse.emf.common.util.URI
 import org.eclipse.emf.ecore.resource.Resource
 import org.eclipse.emf.ecore.resource.impl.ResourceSetImpl
 import org.eclipse.viatra.query.patternlanguage.emf.EMFPatternLanguageStandaloneSetup
 import org.eclipse.viatra.query.patternlanguage.emf.vql.PatternLanguagePackage
-import org.eclipse.xtext.diagnostics.Severity
 import org.eclipse.xtext.resource.XtextResource
-import org.eclipse.xtext.util.CancelIndicator
-import org.eclipse.xtext.validation.CheckMode
 import java.io.File
-
-fun File.walkFiles() = walkTopDown().filter { it.isFile }
 
 fun prepareOxsts() {
     OxstsStandaloneSetup.doSetup()
@@ -52,10 +47,7 @@ class OxstsReader(
 
     fun readFile(file: File): Resource {
         logger.info { "Reading file: $file" }
-        val resource = resourceSet.getResource(URI.createFileURI(file.path), true)
-        resource.load(emptyMap<Any, Any>())
-
-        return resource
+        return resourceSet.getResource(URI.createFileURI(file.path), true)
     }
 
     fun readDirectory(modelDirectory: String) {
@@ -79,7 +71,7 @@ class OxstsReader(
             userResources += readFile(file)
         }
 
-        validateResources()
+        validateAndLoadResources()
     }
 
     fun readModel(modelPath: String) {
@@ -102,49 +94,14 @@ class OxstsReader(
 
         userResources += readFile(modelFile)
 
-        validateResources()
+        validateAndLoadResources()
     }
 
-    private fun validateResources() {
+    private fun validateAndLoadResources() {
         logger.info("Validating resources")
 
         for (resource in resourceSet.resources) {
-            if (resource.errors.any()) {
-                logger.error { "Errors found in file (${resource.uri.toFileString()})" }
-
-                for (error in resource.errors) {
-                    logger.error(error.message)
-                }
-
-                error("Errors found in file (${resource.uri.toFileString()})}")
-            }
-            if (resource.warnings.any()) {
-                logger.warn { "Warnings found in file (${resource.uri.toFileString()})" }
-
-                for (warning in resource.warnings) {
-                    logger.warn(warning.message)
-                }
-            }
-            val validator = (resource as XtextResource).resourceServiceProvider.resourceValidator
-            val issues = validator.validate(resource, CheckMode.ALL, CancelIndicator.NullImpl)
-            if (issues.any()) {
-                logger.info { "Issues found in file (${resource.uri.toFileString()})" }
-
-                for (issue in issues) {
-                    when (issue.severity) {
-                        Severity.INFO -> {
-                            logger.info { "${issue.uriToProblem.toFileString()}[${issue.lineNumber}:${issue.column}] ${issue.message}" }
-                        }
-                        Severity.WARNING -> {
-                            logger.warn { "${issue.uriToProblem.toFileString()}[${issue.lineNumber}:${issue.column}] ${issue.message}" }
-                        }
-                        Severity.ERROR -> {
-                            logger.error { "${issue.uriToProblem.toFileString()}[${issue.lineNumber}:${issue.column}] ${issue.message}" }
-                        }
-                        else -> { }
-                    }
-                }
-            }
+            XtextResourceValidator.validateAndLoadResource(resource)
         }
 
         logger.info("Validation successful!")
