+++ subprojects/semantifyr/src/main/kotlin/utils/SystemUtils.kt
@@ -1,5 +1,12 @@
+/*
+ * SPDX-FileCopyrightText: 2025 The Semantifyr Authors
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+
 package hu.bme.mit.semantifyr.oxsts.semantifyr.utils
 
+import java.io.File
 import kotlin.reflect.KProperty
 
 open class EnvVar {
@@ -12,3 +19,5 @@ open class EnvVar {
 
 @Suppress("ClassName")
 object environment : EnvVar()
+
+fun File.walkFiles() = walkTopDown().filter { it.isFile }
